requirements.txt
fastapi
pymongo
pydantic
uvicorn
python-docx
wget
starlette
openpyxl
img2pdf
pymorphy3
petrovich
python-dotenv
argon2-cffi